refactor : answer comment hooks
