+++ frontend/hooks/react-query/answer/useGetAnswer.tsx
@@ -9,13 +9,23 @@ type AnswerData = {
   pageInfo: PageInfo;
 };
 
-export const useProjectAnswer = () => {
+type Props = {
+  category: 'PROJECT' | 'ARTICLE';
+};
+
+export const useGetAnswer = ({ category }: Props) => {
   const router = useRouter();
   const { id } = router.query;
   const { isLoading, error, data, refetch } = useQuery<AnswerData, Error>(
-    ['project-answer-list', id],
+    [`${category}-answer-list`, id],
     async () => {
-      if (!router.route.includes('create')) {
+      if (router.route.includes('create')) {
+        return;
+      }
+      if (category === 'ARTICLE') {
+        return await api(`/articles/${id}/answers`).then((res) => res.data);
+      }
+      if (category === 'PROJECT') {
         return await api(`/projects/${id}/answers`).then((res) => res.data);
       }
     }
